Rename DEMRasterSources and friends to TINRasterSource
